requirements.txt
telebot
os
sqlite3
google.generativeai
datetime
